Major Bug Fix quantization fixed at low volume!
At one point there was old code using int instead of double or float in the frequency muxer. fixed.
begin migrating to double from float to reduce quantization noise.

+++ multiband_compressor/freq_mux/freq_mux.h
@@ -8,17 +8,17 @@ struct fmux{
   afilter* end_ptr;
   int poles;
   int num_segs;
-  float* freqs;
+  double* freqs;
 };
 
 typedef struct fmux* fmux;
 
 fmux create_fmux(int poles,int rate,int max_freq,int bins);
 fmux create_fmux_from_pre(int poles,int rate,int* freq_table,int size);
-float power_at(fmux mux,int location);
-int set_power_at(fmux mux,int location,float value);
-void mux(fmux mux,int input);
-int demux(fmux mux);
+double power_at(fmux mux,int location);
+int set_power_at(fmux mux,int location,double value);
+void mux(fmux mux,double input);
+double demux(fmux mux);
 
 void free_mux(fmux mux);
