+++ multiband_compressor/multi_pole/RC/rc.h
@@ -2,9 +2,9 @@
 #define RC_FILTERS
 
 struct rc_filter_info{
-  float alpha;
-  float prev;
-  float prev_raw;
+  double alpha;
+  double prev;
+  double prev_raw;
   int direction;
 };
 
@@ -12,8 +12,8 @@ struct rc_filter_info{
 typedef struct rc_filter_info rc_filter_info;
 
 //RC low pass and high pass filters
-rc_filter_info* create_rc_filter(float frequency,int rate, int direction);
-float do_rc_filter(rc_filter_info* rcf,float in);
+rc_filter_info* create_rc_filter(double frequency,int rate, int direction);
+double do_rc_filter(rc_filter_info* rcf,double in);
 void free_rc_filter(rc_filter_info* rcf);
 
 #endif // !RC_FILTERS
